refactor: separa rotas de API das rotas WEB
