+++ backend/routes/api/assessment.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const controller = require('../controllers/user');
+const controller = require('../../controllers/assessment');
 
 router.post('/', controller.create);
 router.get('/', controller.retrieveAll);
