+++ backend/routes/api/question.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const controller = require('../controllers/criterion');
+const controller = require('../../controllers/question');
 
 router.post('/', controller.create);
 router.get('/', controller.retrieveAll);
