+++ backend/routes/api/user.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const controller = require('../controllers/assessment');
+const controller = require('../../controllers/user');
 
 router.post('/', controller.create);
 router.get('/', controller.retrieveAll);
